requirements.txt
django>=1.6.0
djangorestframework==3.1.1
mock==1.0.1
